news-app-master/requirements.txt
asgiref==3.3.4
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
django-ninja==0.22.2
django-cors-headers==4.4.0
